requirements.txt
# =============================================================================
# WishlistOps Python Dependencies
# =============================================================================
# 
# Core dependencies for Steam marketing automation platform
# Compatible with Python 3.11+
#
# Install: pip install -r requirements.txt
# =============================================================================

# -------------------------------------------------------------------------
# HTTP & API Clients
# -------------------------------------------------------------------------
requests>=2.31.0          # Synchronous HTTP requests (Steam API)
aiohttp>=3.9.0            # Async HTTP requests (parallel API calls)
httpx>=0.25.0             # Modern HTTP client with HTTP/2 support

# -------------------------------------------------------------------------
# Data Validation & Models
# -------------------------------------------------------------------------
pydantic>=2.5.0           # Type-safe data validation and settings
pydantic-settings>=2.1.0  # Environment-based configuration

# -------------------------------------------------------------------------
# Git Operations
# -------------------------------------------------------------------------
GitPython>=3.1.40         # Git repository interactions

# -------------------------------------------------------------------------
# Image Processing
# -------------------------------------------------------------------------
Pillow>=10.0.0            # Image manipulation and compositing
pillow-heif>=0.13.0       # HEIF/HEIC support (optional)

# -------------------------------------------------------------------------
# Configuration & Environment
# -------------------------------------------------------------------------
python-dotenv>=1.0.0      # Load environment variables from .env
tomli>=2.0.1              # TOML parser (Python <3.11 compatibility)
pyyaml>=6.0.1             # YAML parser for configs

# -------------------------------------------------------------------------
# AI & Machine Learning (Google Gemini)
# -------------------------------------------------------------------------
google-generativeai>=0.3.0  # Google Gemini API client

# -------------------------------------------------------------------------
# Discord Integration
# -------------------------------------------------------------------------
discord-webhook>=1.3.0    # Discord webhook notifications
discord.py>=2.3.0         # Full Discord bot (for approval buttons)

# -------------------------------------------------------------------------
# Utilities
# -------------------------------------------------------------------------
python-dateutil>=2.8.2    # Date/time parsing and manipulation
pytz>=2023.3              # Timezone support
tenacity>=8.2.3           # Retry logic with exponential backoff
ratelimit>=2.2.1          # Rate limiting decorators
filelock>=3.13.0          # Thread-safe file locking for state management

# -------------------------------------------------------------------------
# Testing & Quality (Development only)
# -------------------------------------------------------------------------
pytest>=7.4.0             # Testing framework
pytest-asyncio>=0.21.0    # Async test support
pytest-cov>=4.1.0         # Code coverage reports
pytest-mock>=3.12.0       # Mocking utilities

# Type checking
mypy>=1.7.0               # Static type checker
types-requests>=2.31.0    # Type stubs for requests
types-PyYAML>=6.0.12      # Type stubs for PyYAML

# Linting & Formatting
pylint>=3.0.0             # Code linter
black>=23.11.0            # Code formatter
isort>=5.12.0             # Import sorter
flake8>=6.1.0             # Style guide enforcement

# -------------------------------------------------------------------------
# Documentation (Optional)
# -------------------------------------------------------------------------
mkdocs>=1.5.0             # Documentation generator
mkdocs-material>=9.4.0    # Material theme for MkDocs

# -------------------------------------------------------------------------
# Security
# -------------------------------------------------------------------------
safety>=2.3.0             # Dependency vulnerability scanner
bandit>=1.7.5             # Security issue scanner

# -------------------------------------------------------------------------
# Logging & Monitoring
# -------------------------------------------------------------------------
structlog>=23.2.0         # Structured logging
python-json-logger>=2.0.7 # JSON log formatting

# -------------------------------------------------------------------------
# Performance
# -------------------------------------------------------------------------
orjson>=3.9.0             # Fast JSON serialization
ujson>=5.8.0              # Ultra-fast JSON parser (fallback)

# -------------------------------------------------------------------------
# Steam API Specific
# -------------------------------------------------------------------------
# Note: No official Steam Web API Python library exists
# We use requests/aiohttp directly with custom wrappers

# -------------------------------------------------------------------------
# Version Constraints
# -------------------------------------------------------------------------
# Python >=3.11 required for:
# - tomllib (built-in TOML parser)
# - Improved typing features
# - Better async/await performance
# - ExceptionGroup support
